feat: new quizzes tests completed
