+++ src/components/QuizContainerNew/QuestionsListNew/QuestionListItemNew/Answers/ShortAnswer.test.tsx
@@ -1,0 +1,13 @@
+import { ShortAnswers } from "./ShortAnswers";
+
+import renderWithTheme from "@/__tests__/__helpers__/renderWithTheme";
+import { shortAnswers } from "@/node-lib/curriculum-api-2023/fixtures/quizElements.fixture";
+
+describe("Short Answers", () => {
+  it("renders the answer", () => {
+    const { getByText } = renderWithTheme(
+      <ShortAnswers answers={shortAnswers} questionNumber={1} />
+    );
+    expect(getByText("water")).toBeInTheDocument();
+  });
+});
